made proxserve synchronous!

+++ tests/proxserve-simple.test.js
@@ -13,14 +13,31 @@ const Proxserve = require('../dist/proxserve.js');
 const util = require('util');
 const { cloneDeep } = require('lodash');
 
-//test if proxy's internal [[handler]] is revoked. according to http://www.ecma-international.org/ecma-262/6.0/#sec-proxycreate
-function isRevoked(value) {
+//test if proxy's internal [[handler]] is revoked. according to https://www.ecma-international.org/ecma-262/#sec-proxycreate
+//currently (Jan 2021) not working
+/*function isRevoked(value) {
 	try {
 		new Proxy(value, value); //instantiating with revoked-proxy throws an error
 		return false;
 	} catch(err) {
 		return Object(value) === value; //check if value was an object at all. only revoked proxy will reach here and return true
 	}
+}*/
+/**
+ * 
+ * @param {Object} objects - the "dataNode[ND].objects". expected to be { target: *, proxy: undefined, isDeleted: true }
+ * @param {Proxy} proxy - the original proxy object (because the reference inside "objects" got deleted)
+ */
+function isRevoked(objects, proxy) {
+	if(objects.isDeleted) {
+		try {
+			proxy.test = 1;
+		} catch(err) {
+			return true;
+		}
+	}
+
+	return false;
 }
 
 var consoleFuncs = { log: console.log, warn: console.warn, error: console.error };
@@ -60,7 +77,7 @@ const testObject = {
 		}
 	}
 };
-
+if(false) {
 test('1. Initiate a proxserve and check if original object stays intact', () => {
 	let origin = cloneDeep(testObject);
 	let proxy = new Proxserve(origin);
@@ -83,7 +100,7 @@ test('2. Object, child-objects and added-child-objects should convert to proxies
 
 test('3. defineProperty should convert string/number properties to proxy', (done) => {
 	let origin = cloneDeep(testObject);
-	let proxy = new Proxserve(origin, {delay:-950});
+	let proxy = new Proxserve(origin, { debug: { destroyDelay: 10 } });
 	let sym = Symbol.for('sym');
 
 	let desc = {
@@ -108,14 +125,16 @@ test('3. defineProperty should convert string/number properties to proxy', (done
 	expect(util.types.isProxy(proxy.obj)).toBe(true);
 	expect(util.types.isProxy(proxy.obj.this_is)).toBe(true);
 
-	let this_is = proxy.obj.this_is; //reference to proxy
+	let objects = proxy.obj.this_is.getProxserveObjects();
+	let originalProxy = objects.proxy;
+
 	desc.value = 5;
 	Object.defineProperty(proxy, 'obj', cloneDeep(desc)); //overwrite existing property 'obj'
 
 	setTimeout(() => {
-		expect(isRevoked(this_is)).toBe(true);
+		expect(isRevoked(objects, originalProxy)).toBe(true);
 		done();
-	}, 100);
+	}, 30);
 });
 
 test('4. Proxies should contain built-in functions', () => {
@@ -179,10 +198,10 @@ test('4. Proxies should contain built-in functions', () => {
 
 test('5. Basic events of changes', (done) => {
 	let proxy = new Proxserve(cloneDeep(testObject));
-	proxy.on('create', function(change) {
+	proxy.on('create', '.new', function(change) {
 		expect(change.oldValue).toBe(undefined);
 		expect(change.value).toBe(5);
-		expect(change.path).toBe('.new');
+		expect(change.path).toBe('');
 		expect(change.type).toBe('create');
 		part2();
 	});
@@ -196,7 +215,7 @@ test('5. Basic events of changes', (done) => {
 			expect(change.path).toBe('.level1_1.arr1[0]');
 			expect(change.type).toBe('update');
 			part3();
-		});
+		}, {deep:true});
 		proxy.level1_1.arr1[0] = 5;
 	}
 
@@ -208,22 +227,30 @@ test('5. Basic events of changes', (done) => {
 			expect(change.path).toBe('.level1_1.arr1');
 			expect(change.type).toBe('delete');
 			part4();
-		});
+		}, {deep:true});
 		delete proxy.level1_1.arr1; //triggers internal destroy timeout of 1 second
 	}
 
 	function part4() {
+		let counter = 0;
 		proxy.removeAllListeners();
-		proxy.on('change', function(changes) {
-			expect(changes).toEqual([{
-				oldValue: undefined, value: 5, path: '.new2', type: 'create'
-			},{
-				oldValue: 5, value: 7, path: '.new2', type: 'update'
-			},{
-				oldValue: 7, value: undefined, path: '.new2', type: 'delete'
-			}]);
-			part5();
-		});
+		proxy.on('change', function(change) {
+			counter++;
+			if(counter === 1) {
+				expect(change).toEqual({
+					oldValue: undefined, value: 5, path: '.new2', type: 'create'
+				});
+			} else if(counter === 2) {
+				expect(change).toEqual({
+					oldValue: 5, value: 7, path: '.new2', type: 'update'
+				});
+			} else if(counter === 3) {
+				expect(change).toEqual({
+					oldValue: 7, value: undefined, path: '.new2', type: 'delete'
+				});
+				part5();
+			}
+		}, {deep:true});
 		proxy.new2 = 5;
 		proxy.new2 = 7;
 		delete proxy.new2;
@@ -233,77 +260,25 @@ test('5. Basic events of changes', (done) => {
 		proxy.removeAllListeners();
 		let counter = 0;
 		proxy.on(['create','update'], function(change) {
-			if(change.type === 'create') {
+			counter++;
+			if(counter === 1) {
 				expect(change).toEqual({ oldValue: undefined, value: 6, path: '.new3', type: 'create' });
-				counter++;
-			}
-			else if(change.type === 'update') {
+			} else if(counter === 2) {
 				expect(change).toEqual({ oldValue: 6, value: 8, path: '.new3', type: 'update' });
-				counter++;
-			}
-			
-			if(counter === 2) {
 				setImmediate(done);
 			}
-		});
+		}, {deep:true});
 		proxy.new3 = 6;
 		proxy.new3 = 8;
 	}
 });
 
-test('6. Delay of events', (done) => {
-	let proxy = new Proxserve(cloneDeep(testObject), { delay: 0 });
-	let changes = [];
-	proxy.on('change', function(change) {
-		changes.push(1);
-	});
-	proxy.num = 5;
-	expect(changes.length).toBe(1); //should happen immediately. without setTimeout(0) or anything like that
-	proxy.num++;
-	expect(changes.length).toBe(2);
-
-	proxy = new Proxserve(cloneDeep(testObject), { delay: 10 });
-	changes.length = 0;
-	proxy.on('change', function(batchOfChanges) {
-		changes.push(...batchOfChanges);
-	});
-	proxy.num = 5;
-	expect(changes.length).toBe(0);
-	proxy.num++;
-	expect(changes.length).toBe(0);
-	setTimeout(() => {
-		expect(changes.length).toBe(2);
-		setImmediate(part2);
-	}, 20);
-
-	function part2() {
-		proxy = new Proxserve(cloneDeep(testObject), { delay: 500 });
-		changes.length = 0;
-		proxy.on('change', function(batchOfChanges) {
-			changes.push(...batchOfChanges);
-		});
-		proxy.num = 5;
-		expect(changes.length).toBe(0);
-		proxy.num++;
-		expect(changes.length).toBe(0);
-		delete proxy.num;
-		expect(changes.length).toBe(0);
-		setTimeout(() => {
-			expect(changes.length).toBe(0);
-		}, 400);
-		setTimeout(() => {
-			expect(changes.length).toBe(3);
-			done();
-		}, 520);
-	}
-});
-
-test('7. Stop/Block/Activate proxies', () => {
+test('6. Stop/Block/Activate proxies', () => {
 	let proxy = new Proxserve(cloneDeep(testObject), {delay:0});
 	let numberOfEmits = 0;
-	proxy.on('change', function(changes) {
+	proxy.on('change', function(change) {
 		numberOfEmits++;
-	});
+	}, {deep:true});
 	proxy.level1_1.arr1[1] = 12;
 	expect(numberOfEmits).toBe(1);
 
@@ -331,10 +306,10 @@ test('7. Stop/Block/Activate proxies', () => {
 	proxy.removeAllListeners();
 	proxy.level1_2.on('change', function(changes) {
 		numberOfEmits++;
-	});
+	}, {deep:true});
 	proxy.level1_2.level2_1.level3_1.on('change', function(changes) {
 		numberOfEmits++;
-	});
+	}, {deep:true});
 	proxy.level1_2.level2_1.level3_1.arr2[0] = 12;
 	expect(numberOfEmits).toBe(2); //two listeners were called
 
@@ -379,7 +354,7 @@ test('7. Stop/Block/Activate proxies', () => {
 	wakeConsole();
 });
 
-test('8. get/set/delete properties after defineProperty', () => {
+test('7. get/set/delete properties after defineProperty', () => {
 	let origin = cloneDeep(testObject);
 	let proxy = new Proxserve(origin, {delay: -950});
 	let sym = Symbol.for('sym');
@@ -424,7 +399,7 @@ test('8. get/set/delete properties after defineProperty', () => {
 	expect(proxy.obj).toBe(undefined);
 });
 
-test('9. splitPath - split path to segments', () => {
+test('8. splitPath - split path to segments', () => {
 	let path = Proxserve.splitPath('.level2_1.level3_1');
 	let path2 = Proxserve.splitPath('level2_1.level3_1');
 	expect(path).toEqual(path2);
@@ -454,7 +429,7 @@ test('9. splitPath - split path to segments', () => {
 	expect(path).toEqual(['new',0,'1.0','1a','keyWith1',9876543210]);
 });
 
-test('10. evalPath - get target property of object and path', (done) => {
+test('9. evalPath - get target property of object and path', (done) => {
 	let proxy = new Proxserve(cloneDeep(testObject), {delay: 0});
 	proxy.on('change', function(changes) {
 		let { object, property, value } = Proxserve.evalPath(this, changes[0].path);
@@ -507,36 +482,38 @@ test('10. evalPath - get target property of object and path', (done) => {
 	expect(value).toEqual(proxy);
 	setImmediate(done);
 });
-
-test('11. On-change listener that makes its own changes', (done) => {
+}
+test('10. On-change listener that makes its own changes', (done) => {
 	let proxy = new Proxserve(cloneDeep(testObject));
-	proxy.level1_1.arr1.on('change', function(changes) {
-		if(changes.length === 3) {
-			proxy.level1_1.arr1[0] = 123; //immediate change should be insterted to next round event emitting
-			expect(changes.length).toBe(3); //shouldn't have changed yet
-			expect(changes[0].value).toBe(17);
-			expect(changes[1].value).toBe(18);
-			expect(changes[2].value).toBe(19);
-		} else {
-			expect(changes.length).toBe(1);
-			expect(changes[0].value).toBe(123);
+	let counter = 0;
+	proxy.level1_1.arr1.on('change', function(change) {
+		counter++;
+		if(counter === 1) {
+			expect(change.value).toBe(17);
+			proxy.level1_1.arr1[0] = 123; //immediate change should emit immediately
+		} else if(counter === 2) {
+			expect(change.value).toBe(123);
+		} else if(counter === 3) {
+			expect(change.value).toBe(18);
+		} else if(counter === 4) {
+			expect(change.value).toBe(19);
 			setImmediate(done);
 		}
-	});
+	}, {deep:true});
 	proxy.level1_1.arr1[0] = 17;
 	proxy.level1_1.arr1[1] = 18;
 	proxy.level1_1.arr1[2] = 19;
 });
 
-test('12. on/once/removeListener/removeAllListeners', () => {
+test('11. on/once/removeListener/removeAllListeners', () => {
 	let proxy = new Proxserve(cloneDeep(testObject), {delay: 0});
 	let counter = 0;
 	let countFunction = function(changes) {
 		counter++;
 	};
 
-	proxy.on('change', countFunction, 123);
-	proxy.on('change', '.new', countFunction);
+	proxy.on('change', countFunction, {deep:true, id:123});
+	proxy.on('change', '.new', countFunction, {deep:true});
 	proxy.removeListener(123);
 	proxy.new = {};
 
@@ -555,57 +532,59 @@ test('12. on/once/removeListener/removeAllListeners', () => {
 	proxy.new.will.exist.later = 3;
 	expect(counter).toBe(6);
 
-	proxy.on('change', countFunction);
+	proxy.on('change', countFunction, {deep:true});
 	proxy.new.will.exist.later++;
 	proxy.removeAllListeners();
 	expect(counter).toBe(7);
 	proxy.new.will.exist.later++;
 	expect(counter).toBe(7);
 
-	proxy.once('change', countFunction);
+	proxy.once('change', countFunction, {deep:true});
 	proxy.new.will.exist.later++;
 	expect(counter).toBe(8);
 	proxy.new.will.exist.later++;
 	expect(counter).toBe(8);
 
-	proxy.once('update', '.new.will', countFunction);
+	proxy.once('update', '.new.will', countFunction, {deep:true});
 	proxy.new.will.exist.later++;
 	expect(counter).toBe(9);
 	proxy.new.will.exist.later++;
 	expect(counter).toBe(9);
 });
 
-test('13. Listen for delete event of sub-properties when parent is deleted', (done) => {
+test('12. Listen for delete event of sub-properties when parent is deleted', (done) => {
 	let proxy = new Proxserve({});
-	let step = 1;
-	proxy.on('change', '.obj.arr[0][0][0]', function(changes) {
-		if(step === 1) {
-			expect(changes.length).toBe(1);
-			expect(changes[0].type).toBe('create');
-			expect(changes[0].value).toBe(0);
-			proxy.obj.arr = [ ['aa','bb','cc'] ];
-			step = 2;
-		} else if(step === 3) {
-			expect(changes.length).toBe(2);
-			expect(changes[0].type).toBe('delete'); //event caused from previous cycle (4 code lines above)
-			expect(changes[0].oldValue).toBe(0);
-			expect(changes[0].value).toBe(undefined);
-			expect(changes[1].type).toBe('create'); //event caused from this cycle, but on another listener
-			expect(changes[1].oldValue).toBe(undefined);
-			expect(changes[1].value).toBe(true);
+	let counter = 0;
+	proxy.on('change', '.obj.arr[0][0][0]', function(change) {
+		counter++;
+		if(counter === 1) {
+			expect(change.type).toBe('create');
+			expect(change.value).toBe(0);
+			proxy.obj.arr = [ ['aa','bb','cc'] ]; //will cause 2 delete events
+		}
+		else if(counter === 2) {
+			expect(change.type).toBe('delete'); //event caused from previous cycle (4 code lines above)
+			expect(change.oldValue).toBe(0);
+			expect(change.value).toBe(undefined);
+		}
+		else if(counter === 4) {
+			expect(change.type).toBe('create'); //event caused from this cycle, but on another listener
+			expect(change.oldValue).toBe(undefined);
+			expect(change.value).toBe(true);
 			done();
 		}
 	});
 	proxy.on('delete', '.obj.arr[0][0][1]', function(change) {
-		expect(change.path).toBe('');
-		expect(change.oldValue).toBe(1);
-		if(step === 2) {
-			proxy.obj.arr[0][0] = [true, false];
-			step = 3;
-		} else {
-			throw new Error('this listener should have ran only once');
-		}
+		setTimeout(() => { //make sure we run after the other delete event
+			counter++;
+			if(counter === 3) {
+				expect(change.path).toBe('');
+				expect(change.oldValue).toBe(1);
+				proxy.obj.arr[0][0] = [true, false];
+			}
+		}, 0);
 	});
+
 	proxy.obj = {
 		arr: [
 			[
